+++ src/cmd/set-alias.ts
@@ -1,0 +1,8 @@
+import { store } from '../store'
+import { escapeDots } from '../utils/common'
+import { logger } from '../logger'
+
+export const setAlias = async (name: string, value: string) => {
+  store.set(`alias.${escapeDots(name)}`, value)
+  logger.success(`Added alias '${name}'`)
+}
